requirements.txt
Django~=4.2.4
python-dotenv~=1.0.0
openai~=0.27.8
pyttsx3~=2.90